Revert "[ui_quiz]: Implement create and update quiz form"
