+++ app/forms/viewform/page.tsx
@@ -7,7 +7,7 @@ import { GetQuiz } from "@/common/api/form/quiz.get";
 interface Answer {
     id: string;
     value: string;
-};
+}
 
 interface Question {
     id: number;
@@ -16,26 +16,26 @@ interface Question {
     answers: Answer[];
     results: string[];
     explain: string;
-};
+}
 
 interface Category {
     id: number;
     name: string;
     questions: Question[];
-};
+}
 
 interface GetCategory {
     props?: {
         repo?: Category;
     };
-};
+}
 
 
-export default function Page({ ...props }) {
+export default function Page() {
     const [questions, setQuestions] = useState<Question[] | []>([]);
 
-    const fetchGetQuiz = async (id: any) => {
-        const data = await GetQuiz(id);
+    const fetchGetQuiz = async () => {
+        const data = await GetQuiz();
 
         if (data) {
             console.log(data);
@@ -44,7 +44,7 @@ export default function Page({ ...props }) {
     };
 
     useEffect(() => {
-        fetchGetQuiz('a3833e4f-dc0f-4fe9-9c76-038382dad0e7');
+        fetchGetQuiz();
     }, []);
 
     return (
