+++ package.json
@@ -47,15 +47,13 @@
 		"react-dom": "18.3.1",
 		"recoil": "^0.7.7",
 		"swiper": "^11.1.4",
-		"swr": "^2.2.5",
-		"uuid": "^10.0.0"
+		"swr": "^2.2.5"
 	},
 	"devDependencies": {
 		"@types/js-cookie": "^3.0.6",
 		"@types/node": "20.5.7",
 		"@types/react": "18.3.3",
 		"@types/react-dom": "18.3.0",
-		"@types/uuid": "^10.0.0",
 		"@typescript-eslint/eslint-plugin": "7.2.0",
 		"@typescript-eslint/parser": "7.2.0",
 		"autoprefixer": "10.4.19",
